Update config.js
original setup was initializing Firebase correctly. I kept the initialization of app, auth, and firestore in the same order, which is perfectly fine.

I added the initialization of GoogleAuthProvider (googleProvider) to make sure it’s ready for use if weneed it for authentication with Google.

also, the exports are now clearly defined, making it clear what parts of Firebase are available for use throughout the application.